requirements.txt
requests
PyQt6
PyGithub
pandas